+++ src/main/kotlin/top/potmot/core/database/generate/impl/h2/H2ColumnTypeDefiner.kt
@@ -57,9 +57,9 @@ object H2ColumnTypeDefiner : ColumnTypeDefiner {
             else -> false
         }
 
-    override fun defaultDataSize(typeCode: Int): Long? = null
+    override fun defaultDataSize(typeCode: Int): Int? = null
 
-    override fun defaultNumericPrecision(typeCode: Int): Long? = null
+    override fun defaultNumericPrecision(typeCode: Int): Int? = null
     override fun getTypeName(typeMeta: ColumnTypeMeta): String {
         if (typeMeta.overwriteByRaw) return typeMeta.rawType
